Wave 5b.1: Coordinator interface expansion — 11 tasks complete
Add capability queries, unified execution methods, snapshotSignals,
speed control, clock management, visualization context, slider context,
measurement reading, snapshot management, current resolver context,
and narrowed compiled accessor to SimulationCoordinator.

+++ src/components/passives/potentiometer.ts
@@ -310,7 +310,7 @@ function createPotentiometerElement(
 const POTENTIOMETER_PROPERTY_DEFS: PropertyDefinition[] = [
   {
     key: "resistance",
-    type: PropertyType.INT,
+    type: PropertyType.FLOAT,
     label: "Resistance (Ω)",
     unit: "Ω",
     defaultValue: 10000,
